+++ SeqWiresLib/Processors/fingeredChordsProcessor.cpp
@@ -1,6 +1,6 @@
-#include <SeqWiresLib/Processors/chordsFromNotesProcessor.hpp>
+#include <SeqWiresLib/Processors/fingeredChordsProcessor.hpp>
 #include <SeqWiresLib/Features/trackFeature.hpp>
-#include <SeqWiresLib/Functions/chordsFromNotesFunction.hpp>
+#include <SeqWiresLib/Functions/fingeredChordsFunction.hpp>
 
 #include <BabelWiresLib/Features/rootFeature.hpp>
