+++ Tests/SeqWiresLib/CMakeLists.txt
@@ -1,7 +1,7 @@
 SET( SEQUENCELIB_TESTS_SRCS
       seqWiresLibTests.cpp
       chordMapProcessorTest.cpp
-      chordsFromNotesProcessorTest.cpp
+      fingeredChordsProcessorTest.cpp
       concatenateProcessorTest.cpp
       excerptProcessorTest.cpp
       filteredTrackIteratorTest.cpp
